% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{api_scrape}
\alias{api_scrape}
\title{Scrape multiple scientific grant APIs for relevant projects}
\usage{
api_scrape(query, from, to, sources = c("nsf"))
}
\arguments{
\item{query}{Keyword to search for, single string}

\item{from}{Search beginning date, standard date format}

\item{to}{Search end date, standard date format}

\item{sources}{vector of databases to query. Supported sources: nsf, nih, ies}
}
\value{
A data.frame in wide format
}
\description{
Scrape multiple scientific grant APIs for relevant projects
}
\examples{
api_scrape(query="Qualitative methods", from="1979-01-31", to="1980-01-31", sources=c("nsf", "nih"))
}
