% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{api_scrape_keyword}
\alias{api_scrape_keyword}
\title{Query grant APIs for a single keyword}
\usage{
api_scrape_keyword(query, from, to, sources)
}
\arguments{
\item{query}{Keyword to search for, single string}

\item{from}{Search beginning date, standard date format}

\item{to}{Search end date, standard date format}

\item{sources}{vector of databases to query. Supported sources: nsf, nih, ies}
}
\value{
A data.frame in wide format
}
\description{
Query grant APIs for a single keyword
}
