% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arnold.R
\name{arnold_get}
\alias{arnold_get}
\title{Get Arnold Foundations/Arnold Venture awards}
\usage{
arnold_get(keyword, from_year, to_year)
}
\arguments{
\item{keyword}{Keyword to query, single string}

\item{from_year}{Year to begin search, integer}

\item{to_year}{Year to end search, integer}
}
\value{
a data.frame
}
\description{
Get Arnold Foundations/Arnold Venture awards
}
\examples{
arnold <- arnold_get("qualitative data", 2016, 2017)
}
