% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arnold.R
\name{arnold_standardize}
\alias{arnold_standardize}
\title{Standardize Arnold Foundations/Arnold Venture award results}
\usage{
arnold_standardize(keywords, from_date, to_date)
}
\arguments{
\item{keywords}{Vector of keywords to search}

\item{from_date}{Beginning date object to search}

\item{to_date}{Ending date object to search}
}
\value{
a standardized data.frame
}
\description{
Standardize Arnold Foundations/Arnold Venture award results
}
