% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{award_scrape}
\alias{award_scrape}
\title{Download and query online grant databases}
\usage{
award_scrape(queries, from, to, sources)
}
\arguments{
\item{queries}{Vector of keyword strings to search}

\item{from}{Standard date format to begin search, only year is applied}

\item{to}{Standard date format to end search, only year is applied}

\item{sources}{Vector of sources to search. Supports: NEH, Sloan}
}
\value{
A data.frame
}
\description{
This functions as the glue code between individual API routines and the top level.
After running each individual source routine,
it renames columns in the data.frames so that we can rbind.data.frame() everything together.
Intended to be run through the awardsBot() routine, which handles all input validation.
}
