% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{award_scrape_api}
\alias{award_scrape_api}
\title{Query grant APIs for a single keyword}
\usage{
award_scrape_api(query, from, to, sources)
}
\arguments{
\item{query}{Keyword to search for, single string}

\item{from}{Search beginning date, standard date format}

\item{to}{Search end date, standard date format}

\item{sources}{vector of databases to query. Supported sources: nsf, nih, ies}
}
\value{
A data.frame
}
\description{
Large sources with complex APIs can only support on keyword at a time,
and this is a limitation on the source routines, including NSF and the Federal Reporter.
To remedy this, this function should to be looped in a wrapper for each individual keyword.
Otherwise, this is equivalent in function to award_scrape(),
but providing support to different sources.
}
