% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awardFindR-package.R
\docType{package}
\name{awardFindR-package}
\alias{awardFindR}
\alias{awardFindR-package}
\title{awardFindR: QDR awardFindR}
\description{
Queries a number of scientific awards databases
    Collects relevant results based on keyword and date parameters,
    returns list of projects that fit those criteria.
}
\keyword{internal}
