% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedreporter.R
\name{fedreporter_get}
\alias{fedreporter_get}
\title{Scrape data from the Federal Reporter}
\usage{
fedreporter_get(query, from, to, agency = "usda,dod,nasa,epa")
}
\arguments{
\item{query}{Keyword, single string}

\item{from}{Beginning fiscal year to search, integer}

\item{to}{Ending fiscal year to search, integer}

\item{agency}{Agency keyword for search criteria. Single comma-separated string. Defaults to "usda,dod,nasa,epa".}
}
\value{
A data.frame from the API, or NULL if no results
}
\description{
Query the Federal Reporter API for grants, get the XML response(s) and process it into a data.frame
}
\examples{
nih <- fedreporter_get(query="ethnography", from=2020, to=2021, agency="nih")
}
