% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedreporter.R
\name{fedreporter_stardardize}
\alias{fedreporter_stardardize}
\title{Standardize award results from the Federal Reporter}
\usage{
fedreporter_stardardize(keyword, from_year, to_year)
}
\arguments{
\item{keyword}{Keyword, single string}

\item{from_year}{Beginning fiscal year to search, integer}

\item{to_year}{Ending fiscal year to search, integer}
}
\value{
A standardized data.frame
}
\description{
Standardize award results from the Federal Reporter
}
