% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gates.R
\name{gates_get}
\alias{gates_get}
\title{Query awards from the Bill & Melinda Gates Foundation}
\usage{
gates_get(keyword, from_year, to_year, verbose = FALSE)
}
\arguments{
\item{keyword}{Single keyword to query}

\item{from_year}{Year integer to begin search}

\item{to_year}{Year integer to end search}

\item{verbose}{enable verbose HTTP messages. TRUE/FALSE, default: false}
}
\value{
a data.frame
}
\description{
Query awards from the Bill & Melinda Gates Foundation
}
\examples{
gates <- gates_get("qualitative", 2018, 2020)
}
