% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gates.R
\name{gates_get_details}
\alias{gates_get_details}
\title{Get details from Gates foundation entry}
\usage{
gates_get_details(entry)
}
\arguments{
\item{entry}{Object passed by gates_get}
}
\value{
single row data.frame
}
\description{
Internal use only
}
