% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gates.R
\name{gates_standardize}
\alias{gates_standardize}
\title{Standardize award results from the Bill & Melinda Gates Foundation}
\usage{
gates_standardize(keyword, from_date, to_date)
}
\arguments{
\item{keyword}{Single keyword to query}

\item{from_date}{Date object to begin search}

\item{to_date}{Date object to end search}
}
\value{
a standardized data.frame
}
\description{
Standardize award results from the Bill & Melinda Gates Foundation
}
