% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnegie.R
\name{get_carnegie}
\alias{get_carnegie}
\title{Search Carnegie awards}
\usage{
get_carnegie(keyword, from_year, to_year, verbose = FALSE)
}
\arguments{
\item{keyword}{Keyword to query}

\item{from_year}{Beginning year to search}

\item{to_year}{Ending year to search}

\item{verbose}{enable verbose HTTP messages. TRUE/FALSE, default: false}
}
\value{
a data.frame
}
\description{
Search Carnegie awards
}
\examples{
\dontrun{
carnegie <- get_carnegie("qualitative data", 2016, 2017)
}
}
