% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osociety.R
\name{get_osociety}
\alias{get_osociety}
\title{Query and scrape Open Society foundation awards}
\usage{
get_osociety(keyword, from_year, to_year, verbose = FALSE)
}
\arguments{
\item{keyword}{Keyword to query}

\item{from_year}{Beginning year to search}

\item{to_year}{Ending year to search}

\item{verbose}{enable verbose HTTP messages. TRUE/FALSE, default: false}
}
\value{
a data.frame
}
\description{
Query and scrape Open Society foundation awards
}
\examples{
osociety <- get_osociety("qualitative", 2016, 2019)
}
