% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rockefeller.R
\name{get_rockefeller}
\alias{get_rockefeller}
\title{Search Rockefeller Foundation grants}
\usage{
get_rockefeller(keyword, from_date, to_date, verbose)
}
\arguments{
\item{keyword}{Keyword to query}

\item{from_date}{Date object to begin search}

\item{to_date}{Date object to end search}

\item{verbose}{enable verbose HTTP messages. TRUE/FALSE, default: false}
}
\value{
a data.frame
}
\description{
Search Rockefeller Foundation grants
}
\examples{
\dontrun{
rockefeller <- get_rockefeller("test", "2012-01-01", "2021-01-01")
}
}
