% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsf.R
\name{get_rsf}
\alias{get_rsf}
\title{Get Russell Sage foundation awards}
\usage{
get_rsf(keyword, verbose = FALSE)
}
\arguments{
\item{keyword}{Keyword to query, single string}

\item{verbose}{enable verbose HTTP messages. TRUE/FALSE, default: false}
}
\value{
a data.frame
}
\description{
Get Russell Sage foundation awards
}
\examples{
\dontrun{rsf <- get_rsf("ethnography")}
}
