% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwjf.R
\name{get_rwjf}
\alias{get_rwjf}
\title{Get Robert Wood Johnson foundation awards}
\usage{
get_rwjf(keyword, from_year, to_year, verbose = FALSE)
}
\arguments{
\item{keyword}{Keyword to query}

\item{from_year}{Beginning year to search}

\item{to_year}{Ending year to search}

\item{verbose}{enable verbose HTTP messages. TRUE/FALSE, default: false}
}
\value{
a data.frame
}
\description{
Get Robert Wood Johnson foundation awards
}
\examples{
rwjf <- get_rwjf("qualitative data analysis", 2014, 2014)
}
