% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sloan.R
\name{get_sloan}
\alias{get_sloan}
\title{Search for a set of keywords in the Sloan grants database.}
\usage{
get_sloan(keyword, from_year, to_year, verbose = FALSE, grantee = FALSE)
}
\arguments{
\item{keyword}{Keyword to query}

\item{from_year}{Beginning year to search}

\item{to_year}{Ending year to search}

\item{verbose}{enable verbose HTTP messages. TRUE/FALSE, default: false}

\item{grantee}{Use the web-based search function instead of an internal grep(), ideal for searching names}
}
\value{
A data.frame
}
\description{
Search for a set of keywords in the Sloan grants database.
}
\examples{

\dontrun{
sloan <- get_sloan("case studies", 2018, 2020)
}
}
