% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssrc.R
\name{get_ssrc}
\alias{get_ssrc}
\title{Search SSRC fellowships and grants by keyword and date}
\usage{
get_ssrc(keyword, from_year, to_year, verbose = FALSE)
}
\arguments{
\item{keyword}{Keyword to query}

\item{from_year}{Beginning year to search}

\item{to_year}{Ending year to search}

\item{verbose}{enable verbose HTTP messages. TRUE/FALSE, default: false}
}
\value{
a data.frame
}
\description{
Search SSRC fellowships and grants by keyword and date
}
\examples{
\dontrun{ssrc <- get_ssrc("qualitative", 2015, 2016)}
}
