% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usaspending.R
\name{get_usaspend}
\alias{get_usaspend}
\title{Search the USAspending database}
\usage{
get_usaspend(keywords, from_date, to_date, verbose, payload = NULL)
}
\arguments{
\item{keywords}{Vector of strings to search}

\item{from_date}{Date object to begin search}

\item{to_date}{Date object to end search}

\item{verbose}{enable verbose HTTP messages. TRUE/FALSE, default: false}

\item{payload}{A custom NIH search query object. Only for advanced purposes. default: null}
}
\value{
a data.frame
}
\description{
Search the USAspending database
}
\examples{
\dontrun{
results <- usaspend_get(c("qualitative", "interview"),
 "2019-01-01", "2020-01-01")
}
}
