% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macarthur.R
\name{macarthur_get}
\alias{macarthur_get}
\title{Search MacArthur foundation for awards}
\usage{
macarthur_get(keyword, from_date, to_date, verbose = FALSE)
}
\arguments{
\item{keyword}{Keyword to query, single string}

\item{from_date}{Date object to begin search}

\item{to_date}{Date object to end search}

\item{verbose}{enable verbose HTTP messages. TRUE/FALSE, default: false}
}
\value{
a data.frame
}
\description{
Search MacArthur foundation for awards
}
\examples{
macarthur <- macarthur_get("qualitative",
"1999-01-01", "2020-01-01")
}
