% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macarthur.R
\name{macarthur_standardize}
\alias{macarthur_standardize}
\title{Standardize award results from the MacArthur Foundation}
\usage{
macarthur_standardize(keyword, from_date, to_date)
}
\arguments{
\item{keyword}{Single keyword to query}

\item{from_date}{Date object to begin search}

\item{to_date}{Date object to end search}
}
\value{
a standardized data.frame
}
\description{
Standardize award results from the MacArthur Foundation
}
