% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mellon.R
\name{mellon_get}
\alias{mellon_get}
\title{Search the Andrew W. Mellon Foundation grant database}
\usage{
mellon_get(keyword, from_year, to_year, verbose = FALSE)
}
\arguments{
\item{keyword}{Keyword to query}

\item{from_year}{Beginning year to search}

\item{to_year}{Ending year to search}

\item{verbose}{enable verbose HTTP messages. TRUE/FALSE, default: false}
}
\value{
a data.frame
}
\description{
Search the Andrew W. Mellon Foundation grant database
}
\examples{
mellon <- mellon_get("qualitative", 2013, 2021)
}
