% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mellon.R
\name{mellon_get_details}
\alias{mellon_get_details}
\title{Parse the fields of an entry in the Mellon grant database}
\usage{
mellon_get_details(entry)
}
\arguments{
\item{entry}{XML element}
}
\value{
a single row data.frame
}
\description{
Internal use only
}
