% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mellon.R
\name{mellon_standardize}
\alias{mellon_standardize}
\title{Standardize award results from Mellon}
\usage{
mellon_standardize(keyword, from_year, to_year)
}
\arguments{
\item{keyword}{Keyword to query}

\item{from_year}{Year to begin search}

\item{to_year}{Year to end search}
}
\value{
a standardized data.frame
}
\description{
Standardize award results from Mellon
}
