% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neh.R
\name{neh_query}
\alias{neh_query}
\title{Query a NEH grant data.frame for a keyword}
\usage{
neh_query(query, df, from, to)
}
\arguments{
\item{query}{A single string keyword}

\item{df}{A data.frame of NEH grant data to search through}

\item{from}{Beginning year to search}

\item{to}{Ending year to search}
}
\value{
A data.frame with the relevant results matching the keyword
}
\description{
Query a NEH grant data.frame for a keyword
}
