% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nih.R
\name{nih_get}
\alias{nih_get}
\title{Search NIH RePORTER}
\usage{
nih_get(keyword, from_date, to_date)
}
\arguments{
\item{keyword}{Keyword to query}

\item{from_date}{Date object to begin search}

\item{to_date}{Date object to end search}
}
\value{
a data.frame
}
\description{
Search NIH RePORTER
}
\examples{
nih <- nih_get("ethnography", "2019-01-01", "2019-05-01")
}
