% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nih.R
\name{nih_standardize}
\alias{nih_standardize}
\title{Standardize NIH RePORTER results}
\usage{
nih_standardize(keyword, from_date, to_date)
}
\arguments{
\item{keyword}{Keyword to query}

\item{from_date}{Date object to begin search}

\item{to_date}{Date object to end search}
}
\value{
a standardized data.frame
}
\description{
Standardize NIH RePORTER results
}
