% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsf.R
\name{nsf_get}
\alias{nsf_get}
\title{Search the NSF API for awards}
\usage{
nsf_get(keyword, from_date, to_date, verbose = FALSE, cfda = NULL)
}
\arguments{
\item{keyword}{Keyword to query}

\item{from_date}{Date object to begin search}

\item{to_date}{Date object to end search}

\item{verbose}{enable verbose HTTP messages. TRUE/FALSE, default: false}

\item{cfda}{Comma-separated CFDA codes to include, default: all}
}
\value{
A data.frame
}
\description{
Search the NSF API for awards
}
\examples{
nsf <- nsf_get("ethnography", "2020-01-01", "2020-02-01")
}
