% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsf.R
\name{nsf_standardize}
\alias{nsf_standardize}
\title{Standardize NSF search responses}
\usage{
nsf_standardize(keyword, from_date, to_date)
}
\arguments{
\item{keyword}{Keyword to query, single string}

\item{from_date}{Beginning date object}

\item{to_date}{End date object}
}
\value{
A standardized data.frame
}
\description{
Standardize NSF search responses
}
