% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ophil.R
\name{ophil_details}
\alias{ophil_details}
\title{Extract details from a list in the ophil grants database}
\usage{
ophil_details(entry)
}
\arguments{
\item{entry}{an xml2 node from a nodelist.}
}
\value{
a data.frame with a single entry
}
\description{
Passed internally by ophil_get in a loop. Not intended to be run directly.
}
