% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ophil.R
\name{ophil_get}
\alias{ophil_get}
\title{Grab the Open Philanthropy grants data search for keyword-date combos}
\usage{
ophil_get(keyword, from, to)
}
\arguments{
\item{from}{Beginning year to search}

\item{to}{Ending year to search}

\item{keywords}{Keyword to search for in the project description, single string}
}
\value{
A data.frame
}
\description{
Grab the Open Philanthropy grants data search for keyword-date combos
}
