% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ophil.R
\name{ophil_standardize}
\alias{ophil_standardize}
\title{Standardize the Open Philanthropy grants search}
\usage{
ophil_standardize(keywords, from_date, to_date)
}
\arguments{
\item{keywords}{Vector of keywords to search}

\item{from_date}{Beginning date object to search}

\item{to_date}{Ending date object to search}
}
\value{
a standardized data.frame
}
\description{
Standardize the Open Philanthropy grants search
}
