% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osociety.R
\name{osociety_get}
\alias{osociety_get}
\title{Query and scrape Open Society foundation awards}
\usage{
osociety_get(keyword, from, to)
}
\arguments{
\item{keyword}{Keyword, single string}

\item{from}{Year to begin search, integer}

\item{to}{Year to end search, integer}
}
\value{
a data.frame
}
\description{
Query and scrape Open Society foundation awards
}
\examples{
osociety <- osociety_get("qualitative", 2016, 2019)
}
