% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osociety.R
\name{osociety_standardize}
\alias{osociety_standardize}
\title{Standardize Open Society foundation awards}
\usage{
osociety_standardize(keywords, from_date, to_date)
}
\arguments{
\item{keywords}{Vector of keywords to search}

\item{from_date}{Beginning date object to search}

\item{to_date}{Ending date object to search}
}
\value{
a standardized data.frame
}
\description{
Standardize Open Society foundation awards
}
