% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{request_xml}
\alias{request_xml}
\title{Request XML specifically, instead of potentially, say, json
Sets HTTP Accept header to application/xml}
\usage{
request_xml(url)
}
\arguments{
\item{url}{URL to query}
}
\value{
The HTTP request response, hopefully in XML
}
\description{
Request XML specifically, instead of potentially, say, json
Sets HTTP Accept header to application/xml
}
\examples{
request_xml("http://someapi.org/api?parameters=foo")
}
