% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rockefeller.R
\name{rockefeller_standardize}
\alias{rockefeller_standardize}
\title{Standardize Rockefeller Foundation grants search}
\usage{
rockefeller_standardize(keywords, from_date, to_date)
}
\arguments{
\item{keywords}{Vector of keywords to search}

\item{from_date}{Beginning date object to search}

\item{to_date}{Ending date object to search}
}
\value{
a standardized data.frame
}
\description{
Standardize Rockefeller Foundation grants search
}
