% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsf.R
\name{rsf_get}
\alias{rsf_get}
\title{Get Russell Sage foundation awards}
\usage{
rsf_get(keyword, verbose = FALSE)
}
\arguments{
\item{keyword}{Keyword to query, single string}

\item{verbose}{enable verbose HTTP messages. TRUE/FALSE, default: false}
}
\value{
a data.frame
}
\description{
Get Russell Sage foundation awards
}
\examples{
\dontrun{rsf <- rsf_get("ethnography")}
}
