% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwjf.R
\name{rwjf_standardize}
\alias{rwjf_standardize}
\title{Standardize Robert Wood Johnson foundation award results}
\usage{
rwjf_standardize(keywords, from_date, to_date)
}
\arguments{
\item{keywords}{Vector of keywords to search}

\item{from_date}{Beginning date object to search}

\item{to_date}{Ending date object to search}
}
\value{
a standardized data.frame
}
\description{
Standardize Robert Wood Johnson foundation award results
}
