% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sloan.R
\name{sloan_df}
\alias{sloan_df}
\title{Scrape the Sloan grants database from html to a data.frame}
\usage{
sloan_df()
}
\value{
A data.frame
}
\description{
Scrape the Sloan grants database from html to a data.frame
}
\examples{
sloan <- sloan_df()
}
