% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sloan.R
\name{sloan_df_details}
\alias{sloan_df_details}
\title{Extract the text details from a list object in the Sloan gronts database}
\usage{
sloan_df_details(entry)
}
\arguments{
\item{entry}{an xml2 node from a nodelist.}
}
\value{
a data.frame with a single entry
}
\description{
Internal function to extract text details in the Sloan routine.
Passed internally by sloan_df in a loop. Not intended to be run directly.
}
