% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sloan.R
\name{sloan_get}
\alias{sloan_get}
\title{Search for a set of keywords in the Sloan grants database.}
\usage{
sloan_get(keywords, from_year, to_year, generate = FALSE)
}
\arguments{
\item{keywords}{vector of keywords to query}

\item{from_year}{Beginning year to search}

\item{to_year}{Ending year to search}

\item{generate}{Should we generate a new data.frame from the web? TRUE/FALSE}
}
\value{
A data.frame
}
\description{
Search for a set of keywords in the Sloan grants database.
}
\examples{
\dontrun{sloan <- sloan_get(c("qualitative data", "case studies"), 2018, 2020)}
}
