% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sloan.R
\name{sloan_get_query}
\alias{sloan_get_query}
\title{Query the Sloan grant data.frame for a keyword}
\usage{
sloan_get_query(query, df, from, to)
}
\arguments{
\item{query}{A single string keyword}

\item{df}{A data.frame passed by sloan_df}

\item{from}{Beginning year to search (integer)}

\item{to}{Ending year to search (integer)}
}
\value{
A data.frame with the relevant results matching the keyword
}
\description{
For internal use only
}
