% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sloan.R
\name{sloan_search}
\alias{sloan_search}
\title{Search for a set of keywords in the Sloan grants database.}
\usage{
sloan_search(queries, from, to)
}
\arguments{
\item{queries}{vector of keywords to query}

\item{from}{Beginning year to search}

\item{to}{Ending year to search}
}
\value{
A data.frame with Sloan grants that match the keywords and date range
}
\description{
This function will scrape and format the sloan data from the web
This is the intended way to run the entire set of Sloan routines
}
\examples{
sloan_search(c("qualitative data", "case studies"), 2018, 2020)
}
