% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssrc.R
\name{ssrc_get_details}
\alias{ssrc_get_details}
\title{Extract details from SSRC webpage entry}
\usage{
ssrc_get_details(entry)
}
\arguments{
\item{entry}{XML object passed by ssrc_get()}
}
\value{
A single line of a data.frame
}
\description{
Extract details from SSRC webpage entry
}
