% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssrc.R
\name{ssrc_standardize}
\alias{ssrc_standardize}
\title{Standardize search for SSRC fellowships and grants}
\usage{
ssrc_standardize(keywords, from_date, to_date)
}
\arguments{
\item{keywords}{Vector of keywords to search}

\item{from_date}{Beginning date object to search}

\item{to_date}{Ending date object to search}
}
\value{
a standardized data.frame
}
\description{
Standardize search for SSRC fellowships and grants
}
