% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{static_scrape}
\alias{static_scrape}
\title{Download and query online static grant databases}
\usage{
static_scrape(keywords, from, to, sources = c("neh", "sloan"))
}
\arguments{
\item{keywords}{Vector of keyword strings to search}

\item{from}{Standard date format to begin search, only year is applied}

\item{to}{Standard date format to end search, only year is applied}

\item{sources}{Vector of sources to search. Supports: NEH, Sloan}
}
\value{
A harmonized data.frame
}
\description{
Download and query online static grant databases
}
\examples{
static_scrape(c("qualitative data", "ethnography"),
from="2020-01-31", to="2021-01-31",
sources=c("neh", "sloan"))
}
