% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{substr_right}
\alias{substr_right}
\title{Substring from the right}
\usage{
substr_right(x, n)
}
\arguments{
\item{x}{Character string to substring from}

\item{n}{Characters from the right to include}
}
\value{
Shortened character string
}
\description{
Substring from the right
}
