% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templeton.R
\name{templeton_df}
\alias{templeton_df}
\title{Scrape the Templeton grants database from html to a data.frame}
\usage{
templeton_df()
}
\value{
A data.frame
}
\description{
Scrape the Templeton grants database from html to a data.frame
}
\examples{
\dontrun{templeton <- templeton_df()}
}
