% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templeton.R
\name{templeton_get}
\alias{templeton_get}
\title{Search for a set of keywords in the Templeton grants database.}
\usage{
templeton_get(keywords, from_year, to_year)
}
\arguments{
\item{keywords}{vector of keywords to query}

\item{from_year}{Beginning year to search}

\item{to_year}{Ending year to search}
}
\value{
A data.frame
}
\description{
Search for a set of keywords in the Templeton grants database.
}
\examples{
\dontrun{templeton <- templeton_get(c("qualitative data", "case studies"), 2018, 2020)}
}
