% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{text_hash}
\alias{text_hash}
\title{Return basic text hash
*NOT* cryptographic!! this is a toy hasher!}
\usage{
text_hash(string)
}
\arguments{
\item{string}{A character string}
}
\value{
a integer summary hash
}
\description{
Return basic text hash
*NOT* cryptographic!! this is a toy hasher!
}
