% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{title_case}
\alias{title_case}
\title{Capitalize string to Title Case}
\usage{
title_case(string)
}
\arguments{
\item{string}{a character string of ANY CaSe}
}
\value{
A Title Case Character String
}
\description{
Capitalize string to Title Case
}
\examples{
# Should output "Foo Bar"
\dontrun{title_case("foo BAR")}
}
