% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usaspending.R
\name{usaspend_get}
\alias{usaspend_get}
\title{Query the USAspending database}
\usage{
usaspend_get(queries, from, to)
}
\arguments{
\item{queries}{Vector of keywords to search}

\item{from}{Beginning date object to search}

\item{to}{Ending date object to search}
}
\value{
a data.frame
}
\description{
Query the USAspending database
}
\examples{
# Single keyword
results <- usaspend_get("interview", "2018-01-01", "2020-01-01")

# Multiple keywords
results <- usaspend_get(c("qualitative", "interview"), "2019-01-01", "2020-01-01")
}
