% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usaspending.R
\name{usaspend_standardize}
\alias{usaspend_standardize}
\title{Standardize the USAspending database search}
\usage{
usaspend_standardize(keywords, from_date, to_date)
}
\arguments{
\item{keywords}{Vector of keywords to search}

\item{from_date}{Beginning date object to search}

\item{to_date}{Ending date object to search}
}
\value{
a standardized data.frame
}
\description{
Standardize the USAspending database search
}
