% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xml2list_to_df}
\alias{xml2list_to_df}
\title{xml2 list to data.frame}
\usage{
xml2list_to_df(xml2list)
}
\arguments{
\item{xml2list}{List object, typically the result of xml2::as_list() function}
}
\value{
A data.frame
}
\description{
Turns the result of an xml2::as_list() function into a data.frame.
Employs maximalist merging of columns, adding NAs to any missing values.
One way of replacing the XML::xmlToDataFrame function that no longer exists in xml2
}
\examples{
df <- xml2list_to_df(xml2list=xml2::as_list(xml))
}
