% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arnold.R
\name{arnold_standardize}
\alias{arnold_standardize}
\title{Standardize Arnold Foundations/Arnold Venture award results}
\usage{
arnold_standardize(keyword, from_year, to_year)
}
\arguments{
\item{keyword}{Keyword to query, single string}

\item{from_year}{Year to begin search, integer}

\item{to_year}{Year to end search, integer}
}
\value{
a standardized data.frame
}
\description{
Standardize Arnold Foundations/Arnold Venture award results
}
