% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{awardFindR}
\alias{awardFindR}
\title{awardFindR grant search}
\usage{
awardFindR(
  keywords,
  sources = c("arnold", "carnegie", "fedreporter", "gates", "macarthur", "mellon",
    "neh", "nih", "nsf", "ophil", "osociety", "rockefeller", "rsf", "rwjf", "sloan",
    "ssrc", "templeton", "usaspend"),
  from_date = "2019-01-01",
  to_date = Sys.Date()
)
}
\arguments{
\item{keywords}{Path to keywords csv file (1 term per line) or vector.}

\item{sources}{A vector of sources to pull from. Default: all}

\item{from_date}{A date object to limit the search, defaults to Jan 1 2019}

\item{to_date}{A date object to limit the search, defaults to today}
}
\value{
a data.frame
}
\description{
Query a collection of online grant databases for awards.
These queries can be limited by keyword, source and date terms.
}
\examples{
# Results for "ethnography" from NSF between 1/1 and 2/1 2020
\dontrun{awards <- awardFindR("ethnography", "nsf",
"2020-01-01", "2020-02-01")}

# More intensive queries
\dontrun{
# Specific keywords, all sources:
specific <- awardFindR(keywords=c("ethnography", "case studies"))

# Specific keyword, all sources, specific date range:
five_years <- awardFindR(keywords="qualitative",
from_date="2015-01-01", to_date="2020-01-01")
}
}
