% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{awardsBot}
\alias{awardsBot}
\title{awardsBot QDR routine}
\usage{
awardsBot(keywords = "keywords.csv")
}
\arguments{
\item{keywords}{Path to csv file containing keywords to query. Defaults to keywords.csv}
}
\value{
a data.frame
}
\description{
Performs default QDR award query of all awards databases.
}
\examples{
results <- awardsBot(keywords="keywords.csv")
}
