% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnegie.R
\name{carnegie_get}
\alias{carnegie_get}
\title{Get Carnegie awards}
\usage{
carnegie_get(keyword, from, to)
}
\arguments{
\item{keyword}{Keyword to query, single string}

\item{from}{Year to begin search, integer}

\item{to}{Year to end search, integer}
}
\value{
a data.frame
}
\description{
Get Carnegie awards
}
\examples{
carnegie <- carnegie_get("qualitative data", 2016, 2017)
}
