% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnegie.R
\name{carnegie_get}
\alias{carnegie_get}
\title{Search Carnegie awards}
\usage{
carnegie_get(keyword, from_year, to_year)
}
\arguments{
\item{keyword}{Keyword to query, single string}

\item{from_year}{Year to begin search, integer}

\item{to_year}{Year to end search, integer}
}
\value{
a data.frame
}
\description{
Search Carnegie awards
}
\examples{
\dontrun{
carnegie <- carnegie_get("qualitative data", 2016, 2017)
}
}
