% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedreporter.R
\name{fedreporter_get}
\alias{fedreporter_get}
\title{Search awards from the Federal Reporter}
\usage{
fedreporter_get(
  keyword,
  from_year,
  to_year,
  verbose = FALSE,
  agency = "usda,dod,nasa,epa,ed"
)
}
\arguments{
\item{keyword}{Keyword to query}

\item{from_year}{Beginning year to search}

\item{to_year}{Ending year to search}

\item{verbose}{enable verbose HTTP messages. TRUE/FALSE, default: false}

\item{agency}{Agencies, comma separated. Defaults to "usda,dod,nasa,epa".}
}
\value{
A data.frame
}
\description{
Search awards from the Federal Reporter
}
\examples{
\dontrun{
federal <- fedreporter_get(keyword="ethnography",
from_year=2020, to_year=2021, agency="nih")
}
}
