% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedreporter.R
\name{fedreporter_get}
\alias{fedreporter_get}
\title{Scrape data from the Federal Reporter}
\usage{
fedreporter_get(query, from, to, agency = "nih")
}
\arguments{
\item{query}{Keyword, single string}

\item{from}{Beginning fiscal year to search, integer}

\item{to}{Ending fiscal year to search, integer}

\item{agency}{Agency keyword for search criteria. Single term only. Defaults to "nih".}
}
\value{
A data.frame of raw Federal Reporter data from the API
}
\description{
Query the Federal Reporter API for grants, get the XML response(s) and process it into a data.frame
}
\examples{
\dontrun{nih <- fedreporter_get(query="Qualitative data", from=2018, to=2020, agency="nih")}
}
