% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedreporter.R
\name{fedreporter_get}
\alias{fedreporter_get}
\title{Search awards from the Federal Reporter}
\usage{
fedreporter_get(keyword, from_year, to_year, agency = "usda,dod,nasa,epa,ed")
}
\arguments{
\item{keyword}{Keyword, single string}

\item{from_year}{Beginning fiscal year to search, integer}

\item{to_year}{Ending fiscal year to search, integer}

\item{agency}{Agency keyword for search criteria. Single comma-separated string. Defaults to "usda,dod,nasa,epa".}
}
\value{
A data.frame
}
\description{
Search awards from the Federal Reporter
}
\examples{
federal <- fedreporter_get(keyword="ethnography", from_year=2020, to_year=2021, agency="nih")
}
