% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gates.R
\name{gates_get}
\alias{gates_get}
\title{Query awards from the Bill & Melinda Gates Foundation}
\usage{
gates_get(query, from, to)
}
\arguments{
\item{query}{Single keyword to query}

\item{from}{Date object to begin search}

\item{to}{Date object to end search}
}
\value{
a data.frame
}
\description{
Query awards from the Bill & Melinda Gates Foundation
}
\examples{
gates <- gates_get("qualitative", "2018-01-01", "2020-01-01")
}
