% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macarthur.R
\name{get_macarthur}
\alias{get_macarthur}
\title{Search MacArthur foundation for awards}
\usage{
get_macarthur(keyword, from_year, to_year, verbose = FALSE)
}
\arguments{
\item{keyword}{Keyword to query}

\item{from_year}{Beginning year to search}

\item{to_year}{Ending year to search}

\item{verbose}{enable verbose HTTP messages. TRUE/FALSE, default: false}
}
\value{
a data.frame
}
\description{
Search MacArthur foundation for awards
}
\examples{
macarthur <- get_macarthur("qualitative",
"1999-01-01", "2020-01-01")
}
