% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neh.R
\name{get_neh}
\alias{get_neh}
\title{Get relevant awards from NEH}
\usage{
get_neh(keyword, from_year, to_year, verbose = FALSE)
}
\arguments{
\item{keyword}{Keyword to query}

\item{from_year}{Beginning year to search}

\item{to_year}{Ending year to search}

\item{verbose}{enable verbose HTTP messages. TRUE/FALSE, default: false}

\item{agency}{Agencies, comma separated. Defaults to "usda,dod,nasa,epa".}
}
\value{
A raw data.frame with the relevant results from NEH
}
\description{
Get relevant awards from NEH
}
\examples{
\dontrun{neh <- get_neh("ethnography", 2018, 2020)}
}
