% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nih.R
\name{get_nih}
\alias{get_nih}
\title{Search NIH RePORTER}
\usage{
get_nih(keyword, from_date, to_date, verbose = FALSE, payload = NULL)
}
\arguments{
\item{keyword}{Keyword to query}

\item{from_date}{Date object to begin search}

\item{to_date}{Date object to end search}

\item{verbose}{enable verbose HTTP messages. TRUE/FALSE, default: false}

\item{payload}{A custom NIH search query object. Only for advanced purposes. default: null}
}
\value{
a data.frame
}
\description{
Search NIH RePORTER
}
\examples{
nih <- get_nih("ethnography", "2019-01-01", "2019-05-01")
}
