% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macarthur.R
\name{macarthur_standardize}
\alias{macarthur_standardize}
\title{Standardize award results from the MacArthur Foundation}
\usage{
macarthur_standardize(keywords, from_date, to_date)
}
\arguments{
\item{keywords}{Vector of keywords to search}

\item{from_date}{Beginning date object to search}

\item{to_date}{Ending date object to search}
}
\value{
a standardized data.frame
}
\description{
Standardize award results from the MacArthur Foundation
}
