% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neh.R
\name{neh_query}
\alias{neh_query}
\title{Query a NEH grant data.frame for a keyword}
\usage{
neh_query(query, df)
}
\arguments{
\item{df}{A data.frame of NEH grant data to search through}

\item{keyword}{A single string keyword}
}
\value{
A data.frame with the relevant results matching the keyword
}
\description{
Loop for neh_get(). Not intended to be run directly.
}
