% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neh.R
\name{neh_standardize}
\alias{neh_standardize}
\title{Standardize NEH results}
\usage{
neh_standardize(keywords, from_year, to_year)
}
\arguments{
\item{keywords}{Vector of strings to search for in the project description}

\item{from_year}{Beginning year to search}

\item{to_year}{Ending year to search}
}
\value{
a standardized data.frame
}
\description{
Standardize NEH results
}
