% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nih.R
\name{nih_get}
\alias{nih_get}
\title{Query NIH RePorter API}
\usage{
nih_get(keyword, from, to)
}
\arguments{
\item{keyword}{Keyword to query}

\item{from}{date object to begin search}

\item{to}{date object to end search}
}
\value{
a data.frame of results
}
\description{
Search for keyword-date queries from the NIH RePorter API.
}
\examples{
nih <- nih_get("ethnography", "2019-01-01", "2019-05-01")
}
