% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nih.R
\name{nih_standardize}
\alias{nih_standardize}
\title{Standardize NIH RePORTER results}
\usage{
nih_standardize(keywords, from_date, to_date)
}
\arguments{
\item{keywords}{Vector of keywords to search}

\item{from_date}{Beginning date object to search}

\item{to_date}{Ending date object to search}
}
\value{
a standardized data.frame
}
\description{
Standardize NIH RePORTER results
}
