% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsf.R
\name{nsf_get}
\alias{nsf_get}
\title{Search the NSF API for awards}
\usage{
nsf_get(keyword, from_date, to_date, cfda = "47.076,47.075")
}
\arguments{
\item{keyword}{Keyword to query, single string}

\item{from_date}{Beginning date object}

\item{to_date}{End date object}

\item{cfda}{Comma-separated CFDA codes to include, SBE and EHR only by default}
}
\value{
A data.frame of raw NSF API output, or NULL if no results
}
\description{
Search the NSF API for awards
}
\examples{
nsf <- nsf_get("ethnography", "2020-01-01", "2020-02-01")
}
