% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsf.R
\name{nsf_get}
\alias{nsf_get}
\title{Scrape the NSF api}
\usage{
nsf_get(query, from, to)
}
\arguments{
\item{query}{Keyword to query, single string}

\item{from}{Beginning date, mm/dd/yyyy format}

\item{to}{End date, mm/dd/yyyy format}
}
\value{
A data.frame of raw NSF API output
}
\description{
Query the NSF API for awards, get the XML response(s) and process it into a data.frame
}
\examples{
nsf_get(query="Qualitative methods", from="01/01/1973", to="01/01/2020")
}
