% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsf.R
\name{nsf_get}
\alias{nsf_get}
\title{Scrape the NSF api}
\usage{
nsf_get(query, from, to)
}
\arguments{
\item{query}{Keyword to query, single string}

\item{from}{Beginning date, standard format}

\item{to}{End date, standard format}
}
\value{
A data.frame of raw NSF API output, or NULL if no results
}
\description{
Query the NSF API for awards, get the XML response(s) and process it into a data.frame
}
\examples{
nsf <- nsf_get(query="ethnography", from="2020-01-01", to="2020-02-01")
}
