% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsf.R
\name{nsf_standardize}
\alias{nsf_standardize}
\title{Standardize NSF search responses}
\usage{
nsf_standardize(keywords, from_date, to_date)
}
\arguments{
\item{keywords}{Vector of keywords to search}

\item{from_date}{Beginning date object to search}

\item{to_date}{Ending date object to search}
}
\value{
a standardized data.frame
}
\description{
Standardize NSF search responses
}
