% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ophil.R
\name{ophil_get}
\alias{ophil_get}
\title{Grab the Open Philanthropy grants data search for keyword-date combos}
\usage{
ophil_get(query, from, to)
}
\arguments{
\item{query}{Keyword to search for in the project description, single string}

\item{from}{Beginning year to search}

\item{to}{Ending year to search}
}
\value{
A data.frame
}
\description{
Grab the Open Philanthropy grants data search for keyword-date combos
}
\examples{
ophil <- ophil_get("qualitative", 2019, 2020)
}
