% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ophil.R
\name{ophil_get}
\alias{ophil_get}
\title{Grab the Open Philanthropy grants data search for keyword-date combos}
\usage{
ophil_get(keyword, from_year, to_year, verbose = FALSE)
}
\arguments{
\item{keyword}{Keyword to search for in description, single string}

\item{from_year}{Beginning year to search}

\item{to_year}{Ending year to search}

\item{verbose}{enable verbose HTTP messages. TRUE/FALSE, default: false}
}
\value{
A data.frame
}
\description{
Grab the Open Philanthropy grants data search for keyword-date combos
}
\examples{
ophil <- ophil_get("qualitative", 2019, 2020)
}
