% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ophil.R
\name{ophil_standardize}
\alias{ophil_standardize}
\title{Standardize the Open Philanthropy grants search}
\usage{
ophil_standardize(keyword, from_year, to_year)
}
\arguments{
\item{keyword}{Keyword to search for in the project description, single string}

\item{from_year}{Beginning year to search}

\item{to_year}{Ending year to search}
}
\value{
A data.frame
}
\description{
Standardize the Open Philanthropy grants search
}
