% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osociety.R
\name{osociety_get}
\alias{osociety_get}
\title{Query and scrape Open Society foundation awards}
\usage{
osociety_get(keyword, from_year, to_year, verbose = FALSE)
}
\arguments{
\item{keyword}{Keyword to query}

\item{from_year}{Beginning year to search}

\item{to_year}{Ending year to search}

\item{verbose}{enable verbose HTTP messages. TRUE/FALSE, default: false}
}
\value{
a data.frame
}
\description{
Query and scrape Open Society foundation awards
}
\examples{
osociety <- osociety_get("qualitative", 2016, 2019)
}
