% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osociety.R
\name{osociety_standardize}
\alias{osociety_standardize}
\title{Standardize Open Society foundation awards}
\usage{
osociety_standardize(keyword, from_year, to_year)
}
\arguments{
\item{keyword}{Keyword, single string}

\item{from_year}{Year to begin search, integer}

\item{to_year}{Year to end search, integer}
}
\value{
a data.frame
}
\description{
Standardize Open Society foundation awards
}
