% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{request_xml}
\alias{request_xml}
\title{Request XML specifically}
\usage{
request_xml(url)
}
\arguments{
\item{url}{URL to query}
}
\value{
The HTTP request response, hopefully in XML
}
\description{
Sets HTTP Accept header to application/xml.
This prevents some APIs from returning json by default
}
\examples{
\dontrun{request_xml("http://someapi.org/api.xml?parameters=foo")}
}
