% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rockefeller.R
\name{rockefeller_get}
\alias{rockefeller_get}
\title{Search Rockefeller Foundation grants}
\usage{
rockefeller_get(keyword, from_date, to_date, verbose)
}
\arguments{
\item{keyword}{Single keyword to query}

\item{from_date}{Date object to begin search}

\item{to_date}{Date object to end search}

\item{verbose}{enable verbose HTTP messages. TRUE/FALSE, default: false}
}
\value{
a data.frame
}
\description{
Search Rockefeller Foundation grants
}
\examples{
\dontrun{
rockefeller <- rockefeller_get("test", "2012-01-01", "2021-01-01")
}
}
