% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rockefeller.R
\name{rockefeller_standardize}
\alias{rockefeller_standardize}
\title{Standardize Rockefeller Foundation grants search}
\usage{
rockefeller_standardize(keyword, from_date, to_date)
}
\arguments{
\item{keyword}{Single keyword to query}

\item{from_date}{Date object to begin search}

\item{to_date}{Date object to end search}
}
\value{
a standardized data.frame
}
\description{
Standardize Rockefeller Foundation grants search
}
