% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwjf.R
\name{rwjf_get}
\alias{rwjf_get}
\title{Get Robert Wood Johnson foundation awards}
\usage{
rwjf_get(keyword, from_year, to_year, verbose = FALSE)
}
\arguments{
\item{keyword}{Keyword to query, single string}

\item{from_year}{Year to begin search, integer}

\item{to_year}{Year to end search, integer}

\item{verbose}{enable verbose HTTP messages. TRUE/FALSE, default: false}
}
\value{
a data.frame
}
\description{
Get Robert Wood Johnson foundation awards
}
\examples{
rwjf <- rwjf_get("qualitative data analysis", 2014, 2014)
}
