% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sloan.R
\name{sloan_df}
\alias{sloan_df}
\title{Scrape the Sloan grants database from html to a data.frame}
\usage{
sloan_df(generate = FALSE)
}
\arguments{
\item{generate}{Should we generate a new data.frame from the web? TRUE/FALSE}
}
\value{
A data.frame
}
\description{
Scrape the Sloan grants database from html to a data.frame
}
\examples{
# Make a new version of the sloan data.frame
\dontrun{sloan <- sloan_df(generate=T)}
}
