% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sloan.R
\name{sloan_get}
\alias{sloan_get}
\title{Search for a set of keywords in the Sloan grants database.}
\usage{
sloan_get(keyword, from_year, to_year, verbose = FALSE)
}
\arguments{
\item{keyword}{Keyword to query}

\item{from_year}{Beginning year to search}

\item{to_year}{Ending year to search}

\item{verbose}{enable verbose HTTP messages. TRUE/FALSE, default: false}
}
\value{
A data.frame
}
\description{
Search for a set of keywords in the Sloan grants database.
}
\examples{

\dontrun{
sloan <- sloan_get("case studies", 2018, 2020)
}
}
