% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sloan.R
\name{sloan_get}
\alias{sloan_get}
\title{Search for a set of keywords in the Sloan grants database.}
\usage{
sloan_get(queries, from, to)
}
\arguments{
\item{queries}{vector of keywords to query}

\item{from}{Beginning year to search}

\item{to}{Ending year to search}
}
\value{
A data.frame
}
\description{
Scrape, format and search the Sloan grants database for a series of keywords
}
\examples{
\dontrun{sloan <- sloan_get(c("qualitative data", "case studies"), 2018, 2020)}
}
