% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sloan.R
\name{sloan_standardize}
\alias{sloan_standardize}
\title{Standardize Sloan search}
\usage{
sloan_standardize(keywords, from_year, to_year)
}
\arguments{
\item{keywords}{vector of keywords to query}

\item{from_year}{Beginning year to search}

\item{to_year}{Ending year to search}
}
\value{
A standardized data.frame
}
\description{
Standardize Sloan search
}
