% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssrc.R
\name{ssrc_get}
\alias{ssrc_get}
\title{Search SSRC fellowships and grants by keyword and date}
\usage{
ssrc_get(keyword, from_year, to_year)
}
\arguments{
\item{keyword}{Keyword to query, single string}

\item{from_year}{Year to begin search, integer}

\item{to_year}{Year to end search, integer}
}
\value{
a data.frame
}
\description{
Search SSRC fellowships and grants by keyword and date
}
\examples{
ssrc <- ssrc_get("qualitative", 2015, 2016)
}
