% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssrc.R
\name{ssrc_get}
\alias{ssrc_get}
\title{Search SSRC fellowships and grants by keyword and date}
\usage{
ssrc_get(keyword, from_year, to_year, verbose = FALSE)
}
\arguments{
\item{keyword}{Keyword to query}

\item{from_year}{Beginning year to search}

\item{to_year}{Ending year to search}

\item{verbose}{enable verbose HTTP messages. TRUE/FALSE, default: false}
}
\value{
a data.frame
}
\description{
Search SSRC fellowships and grants by keyword and date
}
\examples{
ssrc <- ssrc_get("qualitative", 2015, 2016)
}
