% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssrc.R
\name{ssrc_get}
\alias{ssrc_get}
\title{Scrape SSRC awards}
\usage{
ssrc_get(keyword, from, to)
}
\arguments{
\item{keyword}{Keyword to query, single string}

\item{from}{Year to begin search, integer}

\item{to}{Year to end search, integer}
}
\value{
a data.frame
}
\description{
Scrape SSRC awards
}
\examples{
ssrc <- ssrc_get("qualitative", 2015, 2016)
}
