% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssrc.R
\name{ssrc_standardize}
\alias{ssrc_standardize}
\title{Standardize search for SSRC fellowships and grants}
\usage{
ssrc_standardize(keyword, from_year, to_year)
}
\arguments{
\item{keyword}{Keyword to query, single string}

\item{from_year}{Year to begin search, integer}

\item{to_year}{Year to end search, integer}
}
\value{
a data.frame
}
\description{
Standardize search for SSRC fellowships and grants
}
