% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{static_scrape}
\alias{static_scrape}
\title{Query static awards databases}
\usage{
static_scrape(keywords, from, to, sources = c("neh", "sloan"))
}
\arguments{
\item{keywords}{Vector of keyword strings to search}

\item{from}{Date object to begin search, only year is applied}

\item{to}{Date object to end search, only year is applied}

\item{sources}{Vector of sources to search. Supports: NEH, Sloan}
}
\value{
A harmonized data.frame
}
\description{
Query static awards databases
}
\examples{
static_scrape(c("qualitative data", "ethnography"), "2020-01-31", "2021-01-31", c("neh", "sloan"))
}
