% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usaspend.R
\name{usaspend_get}
\alias{usaspend_get}
\title{Query the USA Spending API}
\usage{
usaspend_get(queries, from, to)
}
\arguments{
\item{from}{Beginning time period}

\item{to}{Ending time period}

\item{query}{Keyword to query}
}
\value{
a data.frame
}
\description{
Query the USA Spending API
}
