% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usaspending.R
\name{usaspend_get}
\alias{usaspend_get}
\title{Search the USAspending database}
\usage{
usaspend_get(keywords, from_date, to_date, verbose)
}
\arguments{
\item{keywords}{Vector of keywords to search}

\item{from_date}{Beginning date object to search}

\item{to_date}{Ending date object to search}

\item{verbose}{enable verbose HTTP messages. TRUE/FALSE, default: false}
}
\value{
a data.frame
}
\description{
Search the USAspending database
}
\examples{
\dontrun{
results <- usaspend_get(c("qualitative", "interview"),
 "2019-01-01", "2020-01-01")
}
}
